Handle error if machinery-helper cannot be copied/executed

+++ lib/local_system.rb
@@ -152,7 +152,7 @@ EOF
   # Copies a file to the local system
   def inject_file(source, destination)
     FileUtils.copy(source, destination)
-  rescue
+  rescue => e
     raise Machinery::Errors::InjectFileFailed.new(
       "Could not inject file '#{source}' to local system.\n" \
       "Error: #{e}"
@@ -161,7 +161,7 @@ EOF
 
   # Removes a file from the System
   def remove_file(file)
-    File.delete(file)
+    File.delete(file) if File.exist?(file)
   rescue => e
     raise Machinery::Errors::RemoveFileFailed.new(
       "Could not remove file '#{file}' on local system'.\n" \
